% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtuvn.R
\name{dtuvn}
\alias{dtuvn}
\title{Density function of truncated univariate normal distribution}
\usage{
dtuvn(x, mean, sd, lower, upper)
}
\arguments{
\item{x}{value at which density is desired.}

\item{mean}{mean of the underlying univariate normal distribution.}

\item{sd}{standard deviation of the underlying univariate normal distribution.}

\item{lower}{lower bound for truncation.}

\item{upper}{upper bound for truncation.}
}
\value{
\code{dtuvn} returns the density (with same dimension and type as \code{x}) of truncated univariate normal distribution.
}
\description{
\code{dtuvn} calculates the probability density function (pdf) of truncated univariate normal distribution.
}
\examples{
dtuvn(x= -3:3, mean=0, sd=1 ,lower= -2, upper=2)

}
