% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_acc.R
\name{exp_acc_opt}
\alias{exp_acc_opt}
\title{Acceptance rate of translated-exponential rejection sampling}
\usage{
exp_acc_opt(a, b)
}
\arguments{
\item{a}{lower bound for truncation.}

\item{b}{upper bound for truncation.}
}
\description{
\code{exp_acc_opt} calculates the acceptance rate of translated-exponential rejection sampling for the truncation interval (a,b).
}
\examples{
set.seed(1203)
exp_acc_opt(1,2)

}
