% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_rejection.R
\name{exp_rej}
\alias{exp_rej}
\title{Translated-exponential rejection sampling}
\usage{
exp_rej(a, b = Inf, lam = "default")
}
\arguments{
\item{a}{lower bound}

\item{b}{upper bound}

\item{lam}{lambda for translated-exponential only}
}
\value{
\code{exp_rej} returns a list
\code{x}: sampled value; and
\code{acc}: total number of draw used.
}
\description{
\code{exp_rej} is used for translated-exponential rejection sampling.
}
\examples{
set.seed(1)
exp_rej(a=1, b=Inf)

}
