% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfnorm_acc.R
\name{halfnorm_acc}
\alias{halfnorm_acc}
\title{Acceptance rate of half-normal rejection sampling}
\usage{
halfnorm_acc(a, b)
}
\arguments{
\item{a}{lower bound for truncation.}

\item{b}{upper bound for truncation.}
}
\description{
\code{halfnorm_acc} calculates the acceptance rate of half-normal rejection sampling for the truncation interval (a,b).
}
\examples{
set.seed(1203)
halfnorm_acc(1,2)

}
