% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_acc.R
\name{imp_acc}
\alias{imp_acc}
\title{Acceptance rate of truncated univariate normal distribution rejection sampling}
\usage{
imp_acc(a, b)
}
\arguments{
\item{a}{lower bound for truncation.}

\item{b}{upper bound for truncation.}
}
\description{
\code{imp_acc} calculates the acceptance rate of truncated univariate standardized normal distribution rejection sampling for the truncation interval (a,b).
}
\examples{
imp_acc(1,Inf) # Case 1: [a,infty)
imp_acc(-1,1) # Case 2: 0 in [a,b], a<0<b
imp_acc(1,2) # Case 3: [a,b], a>=0

}
