% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_rejection.R
\name{norm_rej}
\alias{norm_rej}
\title{Normal rejection sampling}
\usage{
norm_rej(a, b = Inf)
}
\arguments{
\item{a}{lower bound}

\item{b}{upper bound}
}
\value{
\code{norm_rej} returns a list
\code{x}: sampled value; and
\code{acc}: total number of draw used.
}
\description{
\code{norm_rej} is used for normal rejection sampling.
}
\examples{
set.seed(1)
norm_rej(a=1, b=Inf)

}
