% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_rejection.R
\name{unif_rej}
\alias{unif_rej}
\title{Uniform rejection sampling}
\usage{
unif_rej(a, b)
}
\arguments{
\item{a}{lower bound}

\item{b}{upper bound}
}
\value{
\code{unif_rej} returns a list
\code{x}: sampled value; and
\code{acc}: total number of draw used.
}
\description{
\code{unif_rej} is used for uniform rejection sampling.
}
\examples{
set.seed(1)
unif_rej(a=1, b=2)

}
